CLN: Removing pending Client classes
