--- ibis/backends/base/sql/client.py
@@ -1,7 +1,0 @@
-import abc
-
-from ibis.backends.base import Client
-
-
-class SQLClient(Client, metaclass=abc.ABCMeta):
-    """Generic SQL client."""
