+++ ibis/backends/pandas/execution/temporal.py
@@ -6,7 +6,7 @@ from pandas.core.groupby import SeriesGroupBy
 
 import ibis.expr.datatypes as dt
 import ibis.expr.operations as ops
-from ibis.backends.base import BaseBackend, Client
+from ibis.backends.base import BaseBackend
 from ibis.expr.scope import Scope
 
 from ..core import (
@@ -234,7 +234,7 @@ def execute_timestamp_from_unix(op, data, **kwargs):
 
 
 @pre_execute.register(ops.TimestampNow)
-@pre_execute.register(ops.TimestampNow, (BaseBackend, Client))
+@pre_execute.register(ops.TimestampNow, BaseBackend)
 def pre_execute_timestamp_now(op, *args, **kwargs):
     timecontext = kwargs.get('timecontext', None)
     return Scope({op: pd.Timestamp('now')}, timecontext)
