+++ ibis/expr/rules.py
@@ -291,9 +291,9 @@ def interval(arg, units=None):
 
 @validator
 def client(arg):
-    from ibis.backends.base import BaseBackend, Client
+    from ibis.backends.base import BaseBackend
 
-    return instance_of((Client, BaseBackend), arg)
+    return instance_of(BaseBackend, arg)
 
 
 # ---------------------------------------------------------------------
