+++ ibis/expr/types.py
@@ -200,11 +200,10 @@ class Expr:
 
         Returns
         -------
-        client : Client
+        Backend
             Backend found.
         """
         from ibis.backends.base import BaseBackend
-        from ibis.backends.base.client import Client
 
         seen_backends = set()
 
@@ -218,7 +217,7 @@ class Expr:
                 seen.add(node)
 
                 for arg in node.flat_args():
-                    if isinstance(arg, (Client, BaseBackend)):
+                    if isinstance(arg, BaseBackend):
                         seen_backends.add(arg)
                     elif isinstance(arg, ir.Expr):
                         stack.append(arg.op())
